Remove redundant audio valiables
These settings are already true by default, because the Makefiles check
whether these options are set to false, which they are not. So, no need
to explicitly make them true; as long as they're not false, they'll be
true. Did I just say the same thing twice? I did indeed.